Moving global js error handling into main.js; moving throwClientError to utilities.js
